# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Proxy bridge for LLM API credentials.",
        "title": "LLMVault API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.dev.llmvault.dev",
    "basePath": "/",
    "paths": {
...
        "/v1/credentials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns credentials for the current organization with cursor-based pagination and usage stats.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "List credentials",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by identity ID",
                        "name": "identity_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by identity external ID",
                        "name": "external_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by JSONB metadata (e.g. {\\",
                        "name": "meta",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 50, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor from previous response",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.paginatedResponse-internal_handler_credentialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stores an encrypted LLM API credential for the current organization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Create a credential",
                "parameters": [
                    {
                        "description": "Credential details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.createCredentialRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.credentialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/v1/credentials/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft-deletes a credential by setting its revoked_at timestamp.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Revoke a credential",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.credentialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_useportal_llmvault_internal_model.JSON": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "internal_handler.createCredentialRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "auth_scheme": {
                    "type": "string"
                },
                "base_url": {
                    "type": "string"
                },
                "external_id": {
                    "description": "auto-upserts identity",
                    "type": "string"
                },
                "identity_id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_useportal_llmvault_internal_model.JSON"
                },
                "refill_amount": {
                    "type": "integer"
                },
                "refill_interval": {
                    "type": "string"
                },
                "remaining": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler.credentialResponse": {
            "type": "object",
            "properties": {
                "auth_scheme": {
                    "type": "string"
                },
                "base_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identity_id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "last_used_at": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_useportal_llmvault_internal_model.JSON"
                },
                "provider_id": {
                    "type": "string"
                },
                "refill_amount": {
                    "type": "integer"
                },
                "refill_interval": {
                    "type": "string"
                },
                "remaining": {
                    "type": "integer"
                },
                "request_count": {
                    "type": "integer"
                },
                "revoked_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.paginatedResponse-internal_handler_credentialResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler.credentialResponse"
                    }
                },
                "has_more": {
                    "type": "boolean"
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Logto JWT bearer token (Management API) or Connect session token (Widget API).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}